requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
argon2-cffi=20.1.0=py36h27cfd23_1
async_generator=1.10=py36h28b3542_0
attrs=20.3.0=pyhd3eb1b0_0
bleach=3.3.0=pyhd3eb1b0_0
ca-certificates=2021.1.19=h06a4308_1
certifi=2020.12.5=py36h06a4308_0
cffi=1.14.5=py36h261ae71_0
chardet=3.0.4=pypi_0
cycler=0.10.0=pypi_0
decorator=4.4.2=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
dgl-cu102=0.5.2=pypi_0
entrypoints=0.3=py36_0
future=0.18.2=pypi_0
idna=2.10=pypi_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
ipykernel=5.3.4=py36h5ca1d4c_0
ipython=6.1.0=py36_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.18.0=py36h06a4308_1
jinja2=2.11.3=pyhd3eb1b0_0
jsonschema=3.2.0=py_2
jupyter_client=6.1.7=py_0
jupyter_core=4.7.1=py36h06a4308_0
jupyterlab_pygments=0.1.2=py_0
kiwisolver=1.3.1=pypi_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.1.0=hdf63c60_0
lmdb=1.1.1=pypi_0
markupsafe=1.1.1=py36h7b6447c_0
matplotlib=3.3.4=pypi_0
mistune=0.8.4=py36h7b6447c_0
msgpack=1.0.2=pypi_0
nb_conda_kernels=2.3.1=py36h06a4308_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.0.7=py36_0
nbformat=5.1.2=pyhd3eb1b0_1
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pyhd3eb1b0_0
networkx=2.5=pypi_0
notebook=6.2.0=py36h06a4308_0
numpy=1.19.2=pypi_0
openssl=1.1.1j=h27cfd23_0
packaging=20.9=pyhd3eb1b0_0
pandas=1.1.5=pypi_0
pandoc=2.11=hb0f4dca_0
pandocfilters=1.4.3=py36h06a4308_1
parso=0.8.1=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=7.2.0=pypi_0
pip=20.2.2=py36_0
prometheus_client=0.9.0=pyhd3eb1b0_0
prompt_toolkit=1.0.15=py36_0
protobuf=3.13.0=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pyarrow=3.0.0=pypi_0
pycparser=2.20=py_2
pygments=2.8.0=pyhd3eb1b0_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyrsistent=0.17.3=py36h7b6447c_0
python=3.6.12=hcff3b4d_2
python-dateutil=2.8.1=pyhd3eb1b0_0
pytz=2021.1=pypi_0
pyzmq=20.0.0=py36h2531618_1
readline=8.0=h7b6447c_0
requests=2.24.0=pypi_0
scipy=1.5.2=pypi_0
seaborn=0.11.1=pypi_0
send2trash=1.5.0=pyhd3eb1b0_1
setuptools=49.6.0=py36_0
simplegeneric=0.8.1=py36_2
six=1.15.0=py36h06a4308_0
sqlite=3.33.0=h62c20be_0
tensorboard-logger=0.1.0=pypi_0
tensorboardx=2.1=pypi_0
terminado=0.9.2=py36h06a4308_0
testpath=0.4.4=pyhd3eb1b0_0
tk=8.6.10=hbc83047_0
torch=1.6.0=pypi_0
torchvision=0.7.0=pypi_0
tornado=6.1=py36h27cfd23_0
tqdm=4.56.2=pypi_0
traitlets=4.3.3=py36_0
urllib3=1.25.10=pypi_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py36_1
wheel=0.35.1=py_0
xz=5.2.5=h7b6447c_0
zeromq=4.3.3=he6710b0_3
zipp=3.4.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
